requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymysql==1.1.2
redis==5.0.1
django-redis==5.4.0
Pillow>=10.4.0
django-allauth==0.54.0
celery==5.3.4
django-celery-beat==2.5.0
python-decouple==3.8
boto3==1.29.7
PyPDF2==3.0.1
python-docx==1.1.0
django-jazzmin==2.6.0
reportlab==4.0.7
requests==2.31.0
sentence-transformers==2.2.2
torch>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

